MacAPI/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
python-multipart==0.0.6
sqlalchemy==2.0.20
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyobjc-framework-Vision==9.2